Third draft of Capstone, all but Unit testing complete

+++ src/main/java/com/techelevator/Candy.java
@@ -1,7 +1,9 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Candy extends Item {
-    public Candy(String name, double price, String location){
+    public Candy(String name, BigDecimal price, String location){
         super(name, price, location);
         super.setDispenseMessage("Munch Munch, Yum!");
     }
