+++ src/main/java/com/techelevator/Chip.java
@@ -1,7 +1,9 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Chip extends Item{
-    public Chip(String name, double price, String location){
+    public Chip(String name, BigDecimal price, String location){
         super(name, price, location);
         super.setDispenseMessage("Crunch Crunch, Yum!");
     }
