+++ src/main/java/com/techelevator/Drink.java
@@ -1,7 +1,9 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Drink extends Item{
-    public Drink(String name, double price, String location){
+    public Drink(String name, BigDecimal price, String location){
         super(name, price, location);
         super.setDispenseMessage("Glug Glug, Yum!");
     }
